+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Trace MySql pool cluster calls.
+
 ## 1.82.0
 - Capture synchronous errors in Lambda functions.
 - Handle ARN correctly when a Lambda function is called via an alias.